moved telemetrymessage types to selfservetypes
